Place settings in seperate json, move jwt related functions to jwt.js and refactor other code for readability.
